requirements.txt
deepgram-sdk==5.3.0
flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.0.1
toml==0.10.2